+++ sdk/healthinsights/Azure.Health.Insights.CancerProfiling/tests/Samples/Sample01_InferCancerProfile.cs
@@ -137,11 +137,11 @@ namespace Azure.Health.Insights.CancerProfiling.Tests.Samples
             var oncoPhenotypeData = new OncoPhenotypeData(new List<PatientRecord> { patient1 }) { Configuration = configuration };
             #endregion
             #region Snippet:HealthInsightsCancerProfilingClientInferCancerProfile
-            OncoPhenotypeResult oncoPhenotypeResult = default;
+            OncoPhenotypeResults oncoResults = default;
             try
             {
-                Operation<OncoPhenotypeResult> operation = client.InferCancerProfile(WaitUntil.Completed, oncoPhenotypeData);
-                oncoPhenotypeResult = operation.Value;
+                Operation<OncoPhenotypeResults> operation = client.InferCancerProfile(WaitUntil.Completed, oncoPhenotypeData);
+                oncoResults = operation.Value;
             }
             catch (Exception ex)
             {
@@ -152,37 +152,25 @@ namespace Azure.Health.Insights.CancerProfiling.Tests.Samples
 
             #region Snippet:HealthInsightsCancerProfilingInferCancerProfileViewResults
             // View operation results
-            if (oncoPhenotypeResult.Status == JobStatus.Succeeded)
+            foreach (OncoPhenotypePatientResult patientResult in oncoResults.Patients)
             {
-                OncoPhenotypeResults oncoResults = oncoPhenotypeResult.Results;
-                foreach (OncoPhenotypePatientResult patientResult in oncoResults.Patients)
+                Console.WriteLine($"\n==== Inferences of Patient {patientResult.Id} ====");
+                foreach (OncoPhenotypeInference oncoInference in patientResult.Inferences)
                 {
-                    Console.WriteLine($"\n==== Inferences of Patient {patientResult.Id} ====");
-                    foreach (OncoPhenotypeInference oncoInference in patientResult.Inferences)
+                    Console.WriteLine($"\n=== Clinical Type: {oncoInference.Type.ToString()}  Value: {oncoInference.Value}   ConfidenceScore: {oncoInference.ConfidenceScore} ===");
+                    foreach (InferenceEvidence evidence in oncoInference.Evidence)
                     {
-                        Console.WriteLine($"\n=== Clinical Type: {oncoInference.Type.ToString()}  Value: {oncoInference.Value}   ConfidenceScore: {oncoInference.ConfidenceScore} ===");
-                        foreach (InferenceEvidence evidence in oncoInference.Evidence)
+                        if (evidence.PatientDataEvidence != null)
                         {
-                            if (evidence.PatientDataEvidence != null)
-                            {
-                                var dataEvidence = evidence.PatientDataEvidence;
-                                Console.WriteLine($"Evidence {dataEvidence.Id} {dataEvidence.Offset} {dataEvidence.Length} {dataEvidence.Text}");
-                            }
-                            if (evidence.PatientInfoEvidence != null)
-                            {
-                                var infoEvidence = evidence.PatientInfoEvidence;
-                                Console.WriteLine($"Evidence {infoEvidence.System} {infoEvidence.Code} {infoEvidence.Name} {infoEvidence.Value}");
-                            }
+                            var dataEvidence = evidence.PatientDataEvidence;
+                            Console.WriteLine($"Evidence {dataEvidence.Id} {dataEvidence.Offset} {dataEvidence.Length} {dataEvidence.Text}");
+                        }
+                        if (evidence.PatientInfoEvidence != null)
+                        {
+                            var infoEvidence = evidence.PatientInfoEvidence;
+                            Console.WriteLine($"Evidence {infoEvidence.System} {infoEvidence.Code} {infoEvidence.Name} {infoEvidence.Value}");
                         }
                     }
-                }
-            }
-            else
-            {
-                IReadOnlyList<ResponseError> oncoErrors = oncoPhenotypeResult.Errors;
-                foreach (ResponseError error in oncoErrors)
-                {
-                    Console.WriteLine($"{error.Code} : {error.Message}");
                 }
             }