+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -1,7 +1,18 @@
 # Release History
 
-## 12.18.0-beta.1 (Unreleased)
+## 12.18.0-beta.2 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 12.18.0-beta.1 (2023-12-05)
 - Added support for service version 2024-02-04.
+- Fixed bug where SAS Directory Depth ("sdd") value 10 or over will add 6 to the sdd value
 
 ## 12.17.1 (2023-11-13)
 - Distributed tracing with `ActivitySource` is stable and no longer requires the [Experimental feature-flag](https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/core/Azure.Core/samples/Diagnostics.md).